requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
aiosqlite==0.19.0

# HTTP Client
httpx==0.26.0
aiohttp==3.9.1

# Text-to-Speech
pyttsx3==2.90
gTTS==2.5.0
edge-tts==6.1.9

# PDF Generation
reportlab==4.0.8
fpdf2==2.7.7

# Data Processing
pandas==2.1.4
tqdm==4.66.1

# Environment Variables
python-dotenv==1.0.0

# Validation
pydantic==2.5.3
pydantic-settings==2.1.0

# Audio Processing
pydub==0.25.1

# Utilities
aiofiles==23.2.1
python-jose==3.3.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0

# Code Quality
black==24.1.0
isort==5.13.2
flake8==7.0.0
mypy==1.8.0
